Fix multi-axis encoders.

+++ DCS_Acquisition/include/DCS_ModuleAcquisition.h
@@ -89,10 +89,9 @@ namespace DCS
         struct DCS_API ClinometerEventData
         {
             // TODO
-            float list_cliX[500];
-            float list_cliY[500];
+            f64 list_cliX[500];
+            f64 list_cliY[500];
             DCS::Timer::Timestamp timestamp;
-            
         };
 
         /**
